edulinksz_flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login==0.6.3
gunicorn==21.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
